backend/requirements.txt
# LangChain and LLM packages
langchain-core>=0.3.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langchain-google-genai>=0.0.5
google-generativeai>=0.8.0

# Database and ORM
pymysql>=1.0.2
sqlalchemy>=2.0.0
alembic>=1.12.0

# Web framework
fastapi>=0.108.0
uvicorn>=0.25.0
python-multipart>=0.0.6

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
pyjwt>=2.8.0

# PDF processing
pypdf>=3.15.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.1.0
uuid>=1.30
python-dotenv>=1.0.0
tqdm>=4.66.1

# API requests
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Redis and Security Dependencies
redis>=5.0.0
slowapi>=0.1.9
cryptography>=41.0.0
aiofiles>=23.2.0


